[Dev] Añadir output al API de comentario

+++ Xapp.API/Controllers/CommentController.cs
@@ -54,6 +54,16 @@ namespace Xapp.API.Controllers
             comment.CreateEntity();
 
             var post = await _db.Posts.FirstOrDefaultAsync(x => x.Id == comment.Id);
+            if (post == null)
+            {
+                var output = new ApiResponse<string>
+                {
+                    StatusCode = 400,
+                    Message = "Error",
+                    Result = "No existe el post"
+                };
+                return BadRequest(output);
+            }
 
             post.Comments.Add(comment);
 
@@ -68,6 +78,17 @@ namespace Xapp.API.Controllers
         {
             var comment = await _db.Comments
                 .FirstOrDefaultAsync(x => x.Id == id);
+
+            if(comment == null)
+            {
+                var output = new ApiResponse<string>
+                {
+                    StatusCode = 400,
+                    Message = "Error",
+                    Result = "No existe el comentario"
+                };
+                return BadRequest(output);
+            }
             _db.Comments.Remove(comment);
             await _db.SaveChangesAsync();
 
@@ -78,6 +99,17 @@ namespace Xapp.API.Controllers
         {
             var comment = await _db.Comments
                 .FirstOrDefaultAsync(x => x.Id == id);
+
+            if(comment == null)
+            {
+                var output = new ApiResponse<string>
+                {
+                    StatusCode = 400,
+                    Message = "Error",
+                    Result = "No existe el comentario"
+                };
+                return BadRequest(output);
+            }
             return Ok(comment);
 
         }
@@ -86,6 +118,17 @@ namespace Xapp.API.Controllers
         {
             var comment = await _db.Comments
                 .FirstOrDefaultAsync(x => x.Id == id);
+
+            if (comment == null)
+            {
+                var output = new ApiResponse<string>
+                {
+                    StatusCode = 400,
+                    Message = "Error",
+                    Result = "No existe el comentario"
+                };
+                return BadRequest(output);
+            }
 
             comment.Content = ncomment.Content;
